refactor: remove admin handlers compat bridge
