+++ rustfs/src/admin/handlers/object_zip_download.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::storage_compat::get_global_region;
+use super::super::storage_compat::get_global_region;
 use crate::admin::router::{ADMIN_OBJECT_ZIP_DOWNLOADS_PATH, AdminOperation, Operation, S3Router};
 use crate::app::context::resolve_object_store_handle;
 use crate::auth::{check_key_valid, get_session_token};
@@ -649,7 +649,7 @@ async fn preflight_zip_items(request: &CreateObjectZipDownloadRequest, items: &[
     Ok(())
 }
 
-fn storage_error_to_s3(err: super::storage_compat::Error) -> s3s::S3Error {
+fn storage_error_to_s3(err: super::super::storage_compat::Error) -> s3s::S3Error {
     ApiError::from(err).into()
 }
