+++ rustfs/src/admin/handlers/quota.rs
@@ -14,9 +14,9 @@
 
 //! Quota admin handlers for HTTP API
 
-use super::storage_compat::metadata_sys::BucketMetadataSys;
-use super::storage_compat::quota::checker::QuotaChecker;
-use super::storage_compat::quota::{BucketQuota, QuotaError, QuotaOperation};
+use super::super::storage_compat::metadata_sys::BucketMetadataSys;
+use super::super::storage_compat::quota::checker::QuotaChecker;
+use super::super::storage_compat::quota::{BucketQuota, QuotaError, QuotaOperation};
 use crate::admin::auth::{validate_admin_request, validate_admin_request_with_bucket};
 use crate::admin::router::{AdminOperation, Operation, S3Router};
 use crate::app::context::{resolve_bucket_metadata_handle, resolve_object_store_handle};
@@ -175,7 +175,7 @@ async fn current_usage_from_context(bucket: &str) -> u64 {
         return 0;
     };
 
-    match super::storage_compat::load_data_usage_from_backend(store).await {
+    match super::super::storage_compat::load_data_usage_from_backend(store).await {
         Ok(data_usage_info) => data_usage_info
             .buckets_usage
             .get(bucket)
